refactor: rename helpers.ts to utils.ts
